% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHull_mixmgfa.R
\name{CHull_mixmgfa}
\alias{CHull_mixmgfa}
\title{Convex Hull procedure for Mixture Multigroup Factor Analysis}
\usage{
CHull_mixmgfa(x)
}
\arguments{
\item{x}{Outputobject$overview (class "mixmgfa") resulting from the mixmgfa function, or a merge of multiple such overview tables (vertically concatenated). The table should have 8 or 9 columns (so excluding or including the scree ratios). The rows don't have to be ordered according to the numbers of clusters.}
}
\description{
Performs the CHull for an overview table or a merge of muliple overview tables (e.g., when adding more numbers of clusters or repeating some analyses with more starts to avoid local maxima). The function automatically selects the most optimal solution for each number of clusters, reorders the rows and recomputes the CHull scree ratios.
}
