% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixmgfa_loadings.R
\name{mixmgfa_loadings}
\alias{mixmgfa_loadings}
\title{Mixture multigroup factor analysis for loading (non-)invariance}
\usage{
mixmgfa_loadings(
  data,
  N_gs,
  nclust,
  nfactors = 1,
  maxiter = 5000,
  start = 1,
  nruns = 50,
  design = 0,
  preselect = 10,
  startpartition
)
}
\arguments{
\item{data}{A matrix of vertically concatenated, group-specific (co)variance matrices; or a matrix containing the vertically concatenated raw data for all groups (first column should not contain group ID).}

\item{N_gs}{Vector containing the sample size for each group (in the same order as they appear in the data).}

\item{nclust}{User-specified number of clusters.}

\item{nfactors}{User-specified number of factors.}

\item{maxiter}{Maximum number of iterations. Increase in case of non-convergence.}

\item{start}{Type of start (start = 1: pre-selected random starts, start = 2: start from a user-specified startpartition).}

\item{nruns}{Number of starts (based on pre-selected random partitions when start = 1).}

\item{design}{For confirmatory factor analysis, matrix (with ncol = nfactors) indicating position of zero loadings with '0' and non-zero loadings with '1'. Leave unspecified for exploratory factor analysis (EFA).
(Using different design matrices for different clusters is currently not supported.)}

\item{preselect}{Percentage of best starts taken in pre-selection (for huge data, increase to speed up startprocedure).}

\item{startpartition}{Partition of groups (vector) to start from (use with start = 2 and nruns = 1).}
}
\value{
Output object (list) with:

$z_gks = cluster memberships of groups (posterior classification probabilities)

$pi_ks = mixing proportions (prior classification probabilities)

$Lambda_ks = cluster-specific loadings, access loadings of cluster k via Lambda_ks[[k]]

$Psi_gs = group-specific unique variances, access loadings of group g via Psi_gs[[g]]

$Phi_gks = group- and cluster-specific factor (co)variances, access (co)variances of group g in cluster k via Phi_gks[[g,k]]

$mu_gs = group-specific means, access means of group g via mu_gs[g,]

$bestloglik = final loglikelihood, loglikelihood of best start

$logliks = loglikelihoods (first column) and number of activated constraints on unique variances (second column) for all starts

$nrpars = number of free parameters, to be used for model selection in combination with bestloglik

$convergence = 2 if best start converged on loglikelihood, 1 if converged on parameter changes, 0 if not converged

$nractivatedconstraints = number of constraints on the unique variances (across groups, for best start) to avoid unique variances approaching zero
}
\description{
Find clusters of groups based on their factor loadings, given a user-specified number of clusters.
Intercepts and residual variances are kept group-specific. Factor (co)variances are group- and cluster-specific. Factor means are zero. See referenced paper for details.
}
\references{
De Roover, K., Vermunt, J. K., & Ceulemans, E. (2020). Mixture multigroup factor analysis for unraveling factor loading noninvariance across many groups. Psychological Methods. Advance online publication.
https://psycnet.apa.org/doi/10.1037/met0000355
}
