% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixmgfa_loadingsinterceptsresiduals.R
\name{mixmgfa_loadingsinterceptsresiduals}
\alias{mixmgfa_loadingsinterceptsresiduals}
\title{Mixture multigroup factor analysis for loading, intercept and residual variance (non-)variance}
\usage{
mixmgfa_loadingsinterceptsresiduals(
  data,
  N_gs,
  nclust,
  nfactors,
  maxiter = 5000,
  start = 1,
  nruns = 25,
  design = 0,
  preselect = 10,
  startpartition
)
}
\arguments{
\item{data}{A matrix of vertically concatenated, group-specific (co)variance matrices; or a matrix containing the vertically concatenated raw data for all groups (first column should not contain group ID).}

\item{N_gs}{Vector containing the sample size for each group (in the same order as they appear in the data).}

\item{nclust}{User-specified number of clusters.}

\item{nfactors}{User-specified number of factors.}

\item{maxiter}{Maximum number of iterations. Increase in case of non-convergence.}

\item{start}{Type of start (start = 1: pre-selected random starts, start = 2: start from a user-specified startpartition).}

\item{nruns}{Number of starts (based on pre-selected random partitions when start = 1).}

\item{design}{For confirmatory factor analysis, matrix (with ncol = nfactors) indicating position of zero loadings with '0' and non-zero loadings with '1'. Leave unspecified for exploratory factor analysis (EFA).
(Using different design matrices for different clusters is currently not supported.)}

\item{preselect}{Percentage of best starts taken in pre-selection (for huge data, increase to speed up startprocedure).}

\item{startpartition}{Partition of groups (vector) to start from (use with start = 2 and nruns = 1).}
}
\value{
Output object (list) with:

$z_gks = cluster memberships of groups (posterior classification probabilities)

$pi_ks= mixing proportions (prior classification probabilities)

$Lambda_ks = invariant loadings, access loadings of cluster k via Lambda_ks[[k]]

$Psi_ks = cluster-specific unique variances, access unique variances of cluster k via Psi_ks[[k]]

$Phi_gks = group- and cluster-specific factor (co)variances, access (co)variances of group g in cluster k via Phi_gks[[g,k]]

$tau_ks = cluster-specific intercepts, access intercepts of cluster k via tau_ks[k,]

$alpha_gks = group- and cluster-specific factor means, access factor means of group g in cluster k via alpha_gks[[g,k]]

$bestloglik = final loglikelihood, loglikelihood of best start

$logliks = loglikelihoods (first column) and number of activated constraints on unique variances (second column) for all starts

$nrpars = number of free parameters, to be used for model selection in combination with bestloglik

$convergence = 2 if best start converged on loglikelihood, 1 if converged on parameter changes, 0 if not converged

$nractivatedconstraints = number of constraints on the unique variances (across groups, for best start) to avoid unique variances approaching zero
}
\description{
Finds clusters of groups based on their loadings, intercepts and residual variances, given a user-specified number of clusters.
Factor (co)variances and factor means are group- and cluster-specific.
}
