% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarymixmgfa.R
\name{summary.mixmgfa}
\alias{summary.mixmgfa}
\title{Creates summary of clustering and measurement parameters for estimated MMG-FA model}
\usage{
\method{summary}{mixmgfa}(x, nclust = 0, ...)
}
\arguments{
\item{x}{Outputobject$MMGFAsolutions (class "mixmgfa") resulting from mixmgfa function}

\item{nclust}{Number of clusters you want to create the summary for}
}
\description{
Creates summary of clustering and measurement parameters for estimated MMG-FA model
}
